fix: update junit testing to accomodate new or removed methods
