+++ build.gradle
@@ -44,7 +44,7 @@ test {
 }
 
 application {
-    mainClass.set("scribbles.Scribbles")
+    mainClass.set("scribbles.Launcher")
     mainClassName = 'scribbles.Launcher'
 }
